backend/requirements.txt
fastapi==0.104.1
uvicorn[standard]==0.24.0
motor==3.3.2
pymongo==4.6.0
python-jose[cryptography]==3.3.0
passlib[bcrypt]==1.7.4
python-multipart==0.0.6
pydantic==2.5.0
pydantic-settings==2.1.0
python-dotenv==1.0.0
email-validator==2.1.0
boto3==1.34.0
botocore==1.34.0
argon2-cffi==23.1.0
passlib[argon2]==1.7.4
requests==2.31.0
httpx>=0.24.0
openpyxl>=3.1.0
PyJWT==2.8.0